refactor(shared): move SkillCategory type to shared/types.ts
SkillCategory was defined in registry/classifier.ts but imported back
by shared/types.ts, discover.ts, and build/router.ts — an inverted
dependency (shared/ should be the base layer that other modules import
from, not the other way around).

Moved the type alone to shared/types.ts, where SkillMetadata/Registry
already reference it. classifier.ts, discover.ts, and router.ts now
import it from there. CategoryMeta/CATEGORY_META/EXPLICIT/KEYWORD_RULES
stay in classifier.ts — they're classification business logic/data,
not a shared type, so colocating them with classifySkill() is correct
per this repo's domain-folder convention (not everything scattered
outside shared/ needs to move there).

Also updated shared/schema.ts's stale cross-reference comment
('Must stay in sync with SkillCategory in classifier.ts' -> shared/types.ts).

Verified: typecheck, test (436/436), lint, validate (146/146), sync
all pass.
